+++ src/App.css
@@ -1,62 +1,13 @@
-.container-all {
+.container-background {
+    display: flex;
     height: 100%;
     width: 100%;
+    align-items: center;
     background-size: cover;
     background-position: center;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    align-items: center;
+    justify-content: center;
+    
     background: #FFEFBA;  /* fallback for old browsers */
-background: -webkit-linear-gradient(to bottom, #FFFFFF, #FFEFBA);  /* Chrome 10-25, Safari 5.1-6 */
-background: linear-gradient(to bottom, #FFFFFF, #FFEFBA); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
-
-    display: flex;
-    justify-content: center;
-    /* background:
-    radial-gradient(circle at 100% 50%, transparent 20%, #00D4F0 21%,
-    #00D4F0 34%, transparent 35%, transparent),
-    radial-gradient(circle at 0% 50%, transparent 20%, #00D4F0 21%, #00D4F0 34%, transparent 35%, transparent) 0 -50px;
-    background-color: #EEEEEE;
-    background-size: 75px 100px; */
+    background: -webkit-linear-gradient(to bottom, #FFFFFF, #FFEFBA);  /* Chrome 10-25, Safari 5.1-6 */
+    background: linear-gradient(to bottom, #FFFFFF, #FFEFBA); /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 }
-
-.container-flex {
-    display: flex;
-    margin-left: 20%;
-    margin-right: 20%;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-    width: 100%;
-}
-
-
-.container-grid {
-    width: 300px;
-    display: grid;
-    grid-template-columns: repeat(auto-fill, 70px);
-    justify-content: space-between;
-    grid-gap: 10px;
-}
-
-/* .container-grid {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    gap: 10px;
-} */
-
-/* .container-grid::after {
-    content: "";
-    flex: 1 1 auto;
-  } */
-
-/* .container-grid {
-    display: grid;
-    grid-template-columns: repeat(3, 80px);
-    grid-gap: 20px;
-} */
