+++ src/styles/EndGameModal.style.css
@@ -1,25 +1,146 @@
 .overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  /* background-color: rgba(0, 0, 0, 0.5); */
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.modal-end {
+  height: 180px;
+  width: 270px;
+  align-items: center;
+  background-color: white;
+  padding: 20px;
+  border-radius: 20px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+  animation: popupAnim 0.5s ease-out;
+}
+
+@keyframes popupAnim {
+  0% {
+      transform: translateY(250%);
   }
-  
-  .modal {
-    height: 200px;
-    width: 300px;
-    align-items: center;
-    background-color: white;
-    padding: 20px;
-    border-radius: 20px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+  100% {
+      transform: translateY(0%);
   }
+}
+
+.close {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  cursor: pointer;
+}
+
+.form-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+}
+
+.form-container label {
+  margin-bottom: 10px;
+}
+
+.no-conform {
+  color: rgb(255, 6, 6);
+  font-size: 12px;
+  margin-top: 3px;
+}
+
+.button {
+  margin-top: 10px;
+  align-items: center;
+  appearance: none;
+  background-color: #3EB2FD;
+  background-image: linear-gradient(1deg, #4F58FD, #149BF3 99%);
+  background-size: calc(100% + 20px) calc(100% + 20px);
+  border-radius: 100px;
+  border-width: 0;
+  box-shadow: none;
+  box-sizing: border-box;
+  color: #FFFFFF;
+  cursor: pointer;
+  display: inline-flex;
+  font-family: CircularStd,sans-serif;
+  font-size: 1rem;
+  height: auto;
+  justify-content: center;
+  line-height: 1.5;
+  padding: 6px 20px;
+  position: relative;
+  text-align: center;
+  text-decoration: none;
+  transition: background-color .2s,background-position .2s;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+  vertical-align: top;
+  white-space: nowrap;
+}
+
+.button:active,
+.button:focus {
+  outline: none;
+}
+
+.button:hover {
+  background-position: -20px -20px;
+}
+
+.button:focus:not(:active) {
+  box-shadow: rgba(40, 170, 255, 0.25) 0 0 0 .125em;
+}
+
+.input-field {
+font-size: 14px;
+border-radius: 6px;
+line-height: 1.5;
+padding: 5px 10px;
+transition: box-shadow 100ms ease-in, border 100ms ease-in, background-color 100ms ease-in;
+border: 2px solid #dee1e2;
+background: #dee1e2;
+display: block;
+height: 36px;
+width: 90%;
+:hover {
+    border-color: #ccc;
+}
+:focus{
+    border-color: #4F58FD;
+    background: #fff;
+}
+}
+
+.label-form {
+font-size: 16px;
+}
+
+@media screen and (max-width: 320px) {
+.modal-end {
+  width: 80%;
+  height: 140px;
+}
+.label-form {
+  font-size: 13px;
+}
+.button {
+  padding: 5px 14px;
+  font-size: 0.75rem;
+  margin-top: 8px;
+}
+.no-conform {
+  font-size: 10px;
+}
+}
 
   .txt-container {
     display: flex;
@@ -37,48 +158,4 @@
   .text * {
     margin-bottom: 5px;
     text-align: center;
-  }
-
-  .button {
-    margin-top: 10px;
-    align-items: center;
-    appearance: none;
-    background-color: #3EB2FD;
-    background-image: linear-gradient(1deg, #4F58FD, #149BF3 99%);
-    background-size: calc(100% + 20px) calc(100% + 20px);
-    border-radius: 100px;
-    border-width: 0;
-    box-shadow: none;
-    box-sizing: border-box;
-    color: #FFFFFF;
-    cursor: pointer;
-    display: inline-flex;
-    font-family: CircularStd,sans-serif;
-    font-size: 1rem;
-    height: auto;
-    justify-content: center;
-    line-height: 1.5;
-    padding: 6px 20px;
-    position: relative;
-    text-align: center;
-    text-decoration: none;
-    transition: background-color .2s,background-position .2s;
-    user-select: none;
-    -webkit-user-select: none;
-    touch-action: manipulation;
-    vertical-align: top;
-    white-space: nowrap;
-  }
-  
-  .button:active,
-  .button:focus {
-    outline: none;
-  }
-  
-  .button:hover {
-    background-position: -20px -20px;
-  }
-  
-  .button:focus:not(:active) {
-    box-shadow: rgba(40, 170, 255, 0.25) 0 0 0 .125em;
   }
