+++ src/styles/InputModal.style.css
@@ -4,16 +4,16 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
+    /* background-color: rgba(0, 0, 0, 0.5); */
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
   }
   
-  .modal {
-    height: 200px;
-    width: 300px;
+  .modal-input {
+    height: 180px;
+    width: 270px;
     align-items: center;
     background-color: white;
     padding: 20px;
@@ -110,11 +110,34 @@
   background: #dee1e2;
   display: block;
   height: 36px;
+  width: 90%;
   :hover {
       border-color: #ccc;
   }
   :focus{
       border-color: #4F58FD;
       background: #fff;
+  }
+}
+
+.label-form {
+  font-size: 16px;
+}
+
+@media screen and (max-width: 320px) {
+  .modal-input {
+    width: 80%;
+    height: 140px;
+  }
+  .label-form {
+    font-size: 13px;
+  }
+  .button {
+    padding: 5px 14px;
+    font-size: 0.75rem;
+    margin-top: 8px;
+  }
+  .no-conform {
+    font-size: 10px;
   }
 }
