+++ src/styles/Tile.style.css
@@ -45,3 +45,28 @@
     font-size: 30px;
     user-select: none;
 }
+
+@media screen and (max-width: 300px) {
+    .tile-open {
+        width: 60px;
+        height: 60px;
+    }
+    .tile-closed {
+        width: 60px;
+        height: 60px;
+    }
+}
+
+@media screen and (max-width: 210px) {
+    .tile-open {
+        width: 50px;
+        height: 50px;
+    }
+    .tile-closed {
+        width: 50px;
+        height: 50px;
+    }
+    .emoji {
+        font-size: 20px;
+    }
+}
